Day14

+++ Utilities/Vector.cs
@@ -1,6 +1,6 @@
 ﻿namespace Utilities;
 
-public record struct Vector(int X, int Y)
+public record struct Vector(long X, long Y)
 {
     public static readonly Vector Zero = new(0, 0);
     public static readonly Vector Up = new(0, -1);
